refactor: Move JMS configuration into new file and add event listener implementation to new channels
